mongo_backend/api_requirements.txt
fastapi
uvicorn
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pymongo
python-dotenv
bcrypt==4.0.1